Breaking up classes to be more maintainable.
